experiment/requirements.txt
annotated-types==0.6.0
anthropic==0.16.0
anyio==4.3.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
filelock==3.13.1
fsspec==2024.2.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-aiplatform==1.43.0
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.2
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-generativeai==0.4.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
huggingface-hub==0.21.3
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
mistralai==0.1.3
numpy==1.26.4
openai==1.13.3
orjson==3.9.15
packaging==23.2
pandas==2.2.1
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.3
pydantic-core==2.16.3
pygments==2.17.2
pyreadstat==1.2.6
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pyyaml==6.0.1
requests==2.31.0
rich==13.7.1
rsa==4.9
scipy==1.12.0
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
tenacity==8.2.3
tokenizers==0.15.2
tqdm==4.66.2
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
